+++ src/components/constructor/burger-element/burger-element.jsx
@@ -4,14 +4,16 @@ import { useEffect, useRef } from 'react'
 import { useDrag, useDrop } from 'react-dnd'
 import { useDispatch, useSelector } from 'react-redux'
 
-import { addConstructorItem, moveConstructorItem, setHoveredItemIndex } from '../../../services/constructor/actions'
+import { addConstructorItem, removeConstructorItem, setHoveredItemIndex } from '../../../services/constructor/actions'
 import { selectConstructorItems } from '../../../services/constructor/selectors'
 import { BurgerIngredientType, DragTypes } from '../../../utils/types'
 import BlankConstructorElement from '../blank-constructor-element.jsx/blank-constructor-element'
+import { sortableHoverHandler } from './sortable-hover-handler'
 import styles from './styles.module.css'
 
 const BurgerElement = ({ item, index, arrangement }) => {
   const isBlank = !item
+  const ref = useRef(null)
   const dispatch = useDispatch()
   const items = useSelector(selectConstructorItems)
   const [{ isOver, canDrop }, dropTarget] = useDrop({
@@ -41,60 +43,12 @@ const BurgerElement = ({ item, index, arrangement }) => {
 
   /////////////////////////////////////////////////////////////////
   /////////////////////////////////////////////////////////////////
-  const itemGlob = item
-  const ref = useRef(null)
   const [{ handlerId }, drop] = useDrop({
     accept: DragTypes.FILLINGS_CO,
-    collect(monitor) {
-      return {
-        handlerId: monitor.getHandlerId(),
-      }
-    },
-    hover(item, monitor) {
-      if (!ref.current) {
-        return
-      }
-      const dragIndex = item.index
-      const hoverIndex = index
-      // Don't replace items with themselves
-      if (dragIndex === hoverIndex) {
-        return
-      }
-      // Determine rectangle on screen
-      const hoverBoundingRect = ref.current?.getBoundingClientRect()
-      // Get vertical middle
-      const hoverMiddleY = (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2
-      // Determine mouse position
-      const clientOffset = monitor.getClientOffset()
-      // Get pixels to the top
-      const hoverClientY = clientOffset.y - hoverBoundingRect.top
-
-      // Only perform the move when the mouse has crossed half of the items height
-      // When dragging downwards, only move when the cursor is below 50%
-      // When dragging upwards, only move when the cursor is above 50%
-
-      // Dragging downwards:
-      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-        return
-      }
-      // Dragging upwards:
-      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-        return
-      }
-      // Dragging downwards when item is the last already:
-      if (hoverIndex === 0 && dragIndex === items.length - 1) {
-        return
-      }
-
-      // Time to actually perform the action
-      dispatch(moveConstructorItem(dragIndex, hoverIndex))
-
-      // Note: we're mutating the monitor item here!
-      // Generally it's better to avoid mutations,
-      // but it's good here for the sake of performance
-      // to avoid expensive index searches.
-      item.index = hoverIndex
-    },
+    collect: (monitor) => ({
+      handlerId: monitor.getHandlerId(),
+    }),
+    hover: (item, monitor) => sortableHoverHandler(ref, dispatch, index, items.length, item, monitor),
   })
   const [{ isDragging }, drag] = useDrag({
     type: DragTypes.FILLINGS_CO,
@@ -103,6 +57,7 @@ const BurgerElement = ({ item, index, arrangement }) => {
       isDragging: monitor.isDragging(),
     }),
   })
+
   const opacity = isDragging ? 0 : 1
   const dragStyle = isDragging ? { border: '2px solid red', opacity } : {}
 
@@ -122,9 +77,7 @@ const BurgerElement = ({ item, index, arrangement }) => {
           price={item.price}
           thumbnail={item.image_mobile}
           extraClass={extraClass}
-          handleClose={() => {
-            console.log('close')
-          }}
+          handleClose={() => dispatch(removeConstructorItem(item, index))}
         />
       )}
     </div>
