+++ src/services/constructor/actions.js
@@ -1,9 +1,15 @@
 export const ADD_CONSTRUCTOR_ITEM = 'ADD_CONSTRUCTOR_ITEM'
+export const REMOVE_CONSTRUCTOR_ITEM = 'REMOVE_CONSTRUCTOR_ITEM'
 export const SET_HOVERED_ITEM_INDEX = 'SET_HOVERED_ITEM_INDEX'
 export const MOVE_CONSTRUCTOR_ITEM = 'MOVE_CONSTRUCTOR_ITEM'
 
 export const addConstructorItem = (item, index) => ({
   type: ADD_CONSTRUCTOR_ITEM,
+  payload: { item, index },
+})
+
+export const removeConstructorItem = (item, index) => ({
+  type: REMOVE_CONSTRUCTOR_ITEM,
   payload: { item, index },
 })
